+++ ambra_sdk/__init__.py
@@ -1,13 +1,13 @@
-__version__ = '3.22.3.0-1'
+__version__ = '3.22.4.0-1'
 
 # from footer in api.html
-API_VERSION = 'LBL0022 v54.0 2022-05-25'
+API_VERSION = 'LBL0022 v55.0 2022-08-03'
 
 # equal for API_VERSION based on dictionary.html
 MODELS_VERSION = API_VERSION
 
 # from storage_api.html
-STORAGE_VERSION = 'LBL0038 v22.0 2022-05-18'
+STORAGE_VERSION = 'LBL0038 v23.0 2022-08-03'
 
 DOCS_URL = 'https://dicomgrid.github.io/sdk-python'
 ADDON_DOCS_URL = '{docs_url}/addon.html'.format(docs_url=DOCS_URL)